Refactor core classes
