+++ core/ParamHandler.ts
@@ -1,0 +1,3 @@
+export class ParamHandler{
+
+}
