+++ core/RouteHandler.ts
@@ -1,0 +1,5 @@
+import {ParamHandler} from './ParamHandler';
+
+export class RouteHandler {
+    constructor(private paramHandler: ParamHandler) { }
+}
